requirements.txt
# ===== J.A.R.V.I.S. Dependencies =====

# Async HTTP
httpx>=0.27.0
aiohttp>=3.9.0

# API SDKs
zhipuai>=2.1.0
sniffio>=1.3.0  # implicit dep of zhipuai on Python 3.14
google-generativeai>=0.8.0  # Imagen / Gemini image generation
google-api-python-client>=2.100.0
google-auth-oauthlib>=1.2.0

# Scheduling
apscheduler>=3.10.4

# Telegram
python-telegram-bot>=21.0

# Database
aiosqlite>=0.20.0

# Configuration
python-dotenv>=1.0.1
pyyaml>=6.0.1

# Security
cryptography>=42.0.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-mock>=3.12.0

# Logging
loguru>=0.7.2

# System Monitoring
psutil>=5.9.0

# Utilities
tenacity>=8.2.3
